+++ fabric/src/main/java/io/github/sakurawald/fuji/core/service/style_striper/StyleStriper.java
@@ -12,30 +12,31 @@ import java.util.Map;
 import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import org.jetbrains.annotations.NotNull;
 
 public class StyleStriper {
 
     @SuppressWarnings("RegExpUnnecessaryNonCapturingGroup")
     private static final Pattern TAG_RESOLVER = Pattern.compile("<([^>]+)>");
 
-    private static final Map<String, PermissionDescriptor> CREATED_STYLE_TYPES = new HashMap<>();
+    private static final Map<String, PermissionDescriptor> CREATED_STYLE_KINDS = new HashMap<>();
 
-    public static String stripe(PlayerEntity player, String type, String input) {
-        for (String tag : resolveTags(input)) {
+    public static @NotNull String stripe(@NotNull PlayerEntity player, @NotNull String styleKind, @NotNull String inputString) {
+        for (String tag : resolveTags(inputString)) {
             String tagType = extractTagType(tag);
-            if (!canUseThisTag(player, type, tagType)) {
-                input = input.replace(tag, "");
+            if (!canUseThisTagType(player, styleKind, tagType)) {
+                inputString = inputString.replace(tag, "");
             }
         }
-        return input;
+        return inputString;
     }
 
-    private static String extractTagType(String tag) {
+    private static @NotNull String extractTagType(@NotNull String tag) {
         tag = tag.trim();
 
         if (tag.startsWith("/")) tag = tag.substring(1);
 
-        // remove the escape character for vanilla minecraft sign.
+        // Remove the escape character for vanilla minecraft sign.
         if (tag.endsWith("\\")) tag = tag.substring(0, tag.length() - 1);
 
         int colonIndex = tag.indexOf(':');
@@ -47,9 +48,9 @@ public class StyleStriper {
         return tag;
     }
 
-    private static Set<String> resolveTags(String string) {
-        /* extract tags */
+    private static @NotNull Set<String> resolveTags(@NotNull String string) {
         Set<String> tags = new HashSet<>();
+
         Matcher matcher = TAG_RESOLVER.matcher(string);
         while (matcher.find()) {
             String tag = matcher.group(1);
@@ -62,19 +63,23 @@ public class StyleStriper {
 
     @DocStringProvider(id = 1751999453562L, value = """
         The permission used for `style tags striper`.
-        A player requires the `corresponding permission` to use that `style tag` in `a specific style type`.
+        A player requires the `corresponding permission` to use that `style tag` in `a style kind`.
+
+        For example:
+        1. The `style kind` can be: `sign`, `anvil`...
+        2. The `style tag` can be: `\\<yellow\\>`, `\\<bold\\>`...
         """)
-    private static PermissionDescriptor getOrCreatePermissionDescriptorForStyleType(String styleType) {
-        return CREATED_STYLE_TYPES.computeIfAbsent(styleType, (it) -> {
+    private static @NotNull PermissionDescriptor getOrCreatePermissionDescriptorForStyleKind(@NotNull String styleKind) {
+        return CREATED_STYLE_KINDS.computeIfAbsent(styleKind, (key) -> {
             String pattern = "fuji.style.%s.<style-tag>";
-            pattern = pattern.formatted(styleType);
+            pattern = pattern.formatted(styleKind);
             return new PermissionDescriptor(pattern, 1751999453562L);
         });
     }
 
-    private static boolean canUseThisTag(PlayerEntity player, String type, String tag) {
-        PermissionDescriptor permission = getOrCreatePermissionDescriptorForStyleType(type);
-        return LuckpermsHelper.hasPermission(player.getUuid(), permission, type, tag);
+    private static boolean canUseThisTagType(@NotNull PlayerEntity player, @NotNull String styleKind, @NotNull String tagType) {
+        PermissionDescriptor permission = getOrCreatePermissionDescriptorForStyleKind(styleKind);
+        return LuckpermsHelper.hasPermission(player.getUuid(), permission, styleKind, tagType);
     }
 
 }