+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/color/anvil/ColorAnvilInitializer.java
@@ -7,6 +7,7 @@ import io.github.sakurawald.fuji.core.service.style_striper.StyleStriper;
 import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
 import io.github.sakurawald.fuji.module.initializer.color.anvil.config.model.ColorAnvilConfigModel;
 import net.minecraft.entity.player.PlayerEntity;
+import org.jetbrains.annotations.NotNull;
 
 @Document(id = 1751824946979L, value = """
     This module allows you to use color tags in `anvil` screen.
@@ -15,7 +16,7 @@ public class ColorAnvilInitializer extends ModuleInitializer {
 
     public static final BaseConfigurationHandler<ColorAnvilConfigModel> config = ObjectConfigurationHandler.ofModule(BaseConfigurationHandler.CONFIG_JSON_LITERAL, ColorAnvilConfigModel.class);
 
-    public static String stripeStyleTags(PlayerEntity player, String string) {
+    public static @NotNull String stripeStyleTags(@NotNull PlayerEntity player, @NotNull String string) {
          return StyleStriper.stripe(player, "anvil", string);
     }
 }